requirements.txt
RPi.GPIO==0.7.0
gpiozero===1.5.1
python-dotenv==0.10.1
dependency-injector==3.14.4
pyalsaaudio==0.9.0
luma.led_matrix==1.5.0
gTTS=2.1.1